+++ src/commands/deployDatapackCommand.ts
@@ -1,15 +1,25 @@
 import * as vscode from 'vscode';
 
 import VlocityDatapackService from '../services/vlocityDatapackService';
+import {DatapackCommandOutcome as Outcome, DatapackCommandResult as Result } from '../services/vlocityDatapackService';
 import { DatapackCommand } from './datapackCommand';
 import helper from './commandHelper';
 
 export default class DeployDatapackCommand extends DatapackCommand {
 
-    private progressText : string = 'Deploying Vlocity datapacks...';
+    private repsonseMessages: { [key: number] : (result: Result) => string } = {
+        [Outcome.success]: (r) => `Succesfully deployed ${r.totalCount} datapack(s)`,
+        [Outcome.partial]: (r: Result) => {
+            if (r.errors.length > 0) {
+                return `Unable to deploy all selected datapack(s); deployed ${r.success.length} datapacks with ${r.errors.length} errors`;
+            }
+            return `Unable to deploy all selected datapack(s); deployed ${r.totalCount} out of ${r.totalCount + r.missingCount}`;
+        },
+        [Outcome.error]: (r) => `Failed to deploy the selected datapack(s); see the log for more details`
+    };
 
     constructor(name : string) {
-        super(name, args => this.deployDatapacks(args[1]));
+        super(name, args => this.deployDatapacks(args[1] || [args[0]]));
         
     }
 
@@ -17,31 +27,22 @@ export default class DeployDatapackCommand extends DatapackCommand {
         try {
             
             // prepare input
-            let progressToken = await this.startProgress(this.progressText);
+            let progressToken = await this.startProgress('Deploying Vlocity datapacks...');
             try {
                 var mainfestEntries = await this.resolveManifestEntriesForFiles(selectedFiles);            
                 var result = await this.datapackService.deploy(mainfestEntries);
             } finally {
                 progressToken.complete();
             }
-    
-            // lets do some math to find out how successfull we were
-            let expectedMinResultCount = mainfestEntries.length;
-            let successCount = (result.records || []).reduce((c, r) => c += (r.VlocityDataPackStatus == 'Success') ? 1 : 0, 0);
-            let errorCount = (result.records || []).reduce((c, r) => c += (r.VlocityDataPackStatus != 'Success') ? 1 : 0, 0);
-    
-            // Interpred the results and report back
-            if (successCount > 0 && errorCount > 0) {
-                await helper.showWarningWithRetry(`Unable to deploy all selected datapack(s); deployed ${successCount} datapacks with ${errorCount} errors`, () => this.deployDatapacks(selectedFiles));
-            }  else if (errorCount == 0) {
-                if(successCount >= expectedMinResultCount) {
-                    vscode.window.showInformationMessage(`Succesfully deployed ${successCount} datapack(s)`);
-                } else {
-                    await helper.showWarningWithRetry(`Unable to deploy all selected datapack(s); deployed ${successCount} out of ${expectedMinResultCount}`, () => this.deployDatapacks(selectedFiles));
-                }
-            } else if (successCount == 0) {
-                await helper.showErrorWithRetry(`Failed to deploy the selected datapack(s); see the log for more details`, () => this.deployDatapacks(selectedFiles));
-            }        
+
+            // report UI progress back
+            let message = this.repsonseMessages[result.outcome](result);
+            switch(result.outcome) {
+                case Outcome.success: await vscode.window.showInformationMessage(message); break;
+                case Outcome.partial: await helper.showWarningWithRetry(message, () => this.deployDatapacks(selectedFiles)); break;
+                case Outcome.error: await helper.showErrorWithRetry(message, () => this.deployDatapacks(selectedFiles)); break;
+            }
+
         } catch (err) {
             await helper.showErrorWithRetry(`Error: ${err}`, () => this.deployDatapacks(selectedFiles));
         }