+++ src/commands/refreshDatapackCommand.ts
@@ -1,45 +1,45 @@
 import * as vscode from 'vscode';
 
 import VlocityDatapackService, * as vds from '../services/vlocityDatapackService';
+import {DatapackCommandOutcome as Outcome, DatapackCommandResult as Result } from '../services/vlocityDatapackService';
 import { DatapackCommand } from './datapackCommand';
 import helper from './commandHelper';
 
 export default class RefreshDatapackCommand extends DatapackCommand {
     
-    private progressText : string = 'Refreshing Vlocity datapacks...';
+    private repsonseMessages: { [key: number] : (result: Result) => string } = {
+        [Outcome.success]: (r) => `Succesfully refreshed ${r.totalCount} datapack(s)`,
+        [Outcome.partial]: (r: Result) => {
+            if (r.errors.length > 0) {
+                return `Unable to refresh all selected datapack(s); refreshed ${r.success.length} datapacks with ${r.errors.length} errors`;
+            }
+            return `Unable to refresh all selected datapack(s); refreshed ${r.totalCount} out of ${r.totalCount + r.missingCount}`;
+        },
+        [Outcome.error]: (r) => `Failed to refresh the selected datapack(s); see the log for more details`
+    };
 
     constructor(name : string) {
-        super(name, args => this.refreshDatapacks(args[1]));
+        super(name, args => this.refreshDatapacks(args[1] || [args[0]]));
     }
 
     protected async refreshDatapacks(selectedFiles: vscode.Uri[]) {
         try {
-            // prepare and call
-            let progressToken = await this.startProgress(this.progressText);
+            // call
+            let progressToken = await this.startProgress('Refreshing Vlocity datapacks...');
             try {
                 var datapacks = await this.resolveDatapacksForFiles(selectedFiles);
                 var result = await this.datapackService.export(datapacks, 0);
             } finally {
                 progressToken.complete();
-            }            
-    
-            // lets do some math to find out how successfull we were
-            let expectedMinResultCount = datapacks.length;
-            let successCount = (result.records || []).reduce((c, r) => c += (r.VlocityDataPackStatus == 'Success') ? 1 : 0, 0);
-            let errorCount = (result.records || []).reduce((c, r) => c += (r.VlocityDataPackStatus != 'Success') ? 1 : 0, 0);
-    
-            // Interpred the results and report back
-            if (successCount > 0 && errorCount > 0) {
-                await helper.showWarningWithRetry(`Unable to refresh all selected datapack(s); refreshed ${successCount} datapacks with ${errorCount} errors`, () => this.refreshDatapacks(selectedFiles));
-            }  else if (errorCount == 0) {
-                if(successCount >= expectedMinResultCount) {
-                    vscode.window.showInformationMessage(`Succesfully refreshed ${successCount} datapack(s)`);
-                } else {
-                    await helper.showWarningWithRetry(`Unable to refresh all selected datapack(s); refreshed ${successCount} out of ${expectedMinResultCount}`, () => this.refreshDatapacks(selectedFiles));
-                }
-            } else if (successCount == 0) {
-                await helper.showErrorWithRetry(`Failed to refresh the selected datapack(s); see the log for more details`, () => this.refreshDatapacks(selectedFiles));
-            }        
+            }
+
+            // report UI progress back
+            let message = this.repsonseMessages[result.outcome](result);
+            switch(result.outcome) {
+                case Outcome.success: await vscode.window.showInformationMessage(message); break;
+                case Outcome.partial: await helper.showWarningWithRetry(message, () => this.refreshDatapacks(selectedFiles)); break;
+                case Outcome.error: await helper.showErrorWithRetry(message, () => this.refreshDatapacks(selectedFiles)); break;
+            }
         } catch (err) {
             await helper.showErrorWithRetry(`Error: ${err}`, () => this.refreshDatapacks(selectedFiles));
         }