+++ src/constants.ts
@@ -1,8 +1,9 @@
 /**
  * Default constants used in this extensipn
  */
-export const VERSION = '0.5.0';
+
+export const VERSION = '0.5.2';
 export const CONFIG_SECTION = 'vlocity';
 export const OUTPUT_CHANNEL_NAME = 'Vlocity';
 export const NG_APP_NAME = 'Vlocode';
-export const LOG_DATE_FORMAT = 'hh:mm:ss.SS';
+export const LOG_DATE_FORMAT = 'HH:mm:ss.SS';